fix: Solves container issue
The server was not getting up because husky adds a "prepare" script to package.json ensuring it will be installed before other installations.

Husky is a dev dependency. It evidences that thousands even millions of projects are deployed with their devDependencies together.

I removed prepare script and adjusted the container command and image.